v0.2.1

+++ src/main/java/com/zhong/kisara/service/DataService.java
@@ -9,5 +9,5 @@ import java.util.List;
  * @author zhonghanbo
  */
 public interface DataService {
-    boolean addData(Connection connection, String tableName,List<TableField> tableFieldList, Integer dataSize);
+    boolean addData(Connection connection, String tableName,List<TableField> tableFieldList, Long dataSize);
 }
